+++ protocol/contracts/libraries/LibBytes.sol
@@ -73,8 +73,8 @@ library LibBytes {
         return memBytes;
     }
 
-    function packAddressAndCumulativeStalkPerBDV(address _address, uint96 cumulativeStalkPerBDV) internal pure returns (bytes32){
-        return bytes32(uint256(_address) << 96 | cumulativeStalkPerBDV);
+    function packAddressAndCumulativeStalkPerBDV(address _address, int96 cumulativeStalkPerBDV) internal pure returns (bytes32){
+        return bytes32(uint256(_address) << 96 | uint96(cumulativeStalkPerBDV));
     }
     function getAddressFromBytes(bytes32 data) internal pure returns(address) {
         return address(bytes20(data));
